fix copy-paste of util functions
